pyproject.toml
[project]
name = "mlops-finalProject"
version = "0.1.0"
description = "Continuous Integration using GitHub Actions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "click>=8.3.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "mlflow>=3.6.0",
    "numpy>=2.3.5",
    "onnx>=1.19.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "prometheus-client>=0.23.1",
    "pylint>=4.0.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.2",
]
